Fix remaining environment issues for build.
Use Spack's new ability to save the environment as a pickle.

New command spackdev env <package> [<cmd> [<args>]...] to allow user to operate in the build environment of a particular package.

All references to the build or install areas of a given package to be developed are redirected to the appropriate spackdev directory.

Environments for each package now correctly reflect their configuration.